Add capability to set VLANs on uplink interfaces
Optionally adds vlans to the bridge uplink PF port in addition to the rep port.
If the uplink PF is part of a bond, that interface is used to add the vlans.
By default this feature is disabled and must be enabled in the net-attach-def.

When removing VF this same configuration will remove vlans from the
uplink, but only if no other rep ports are on the bridge with the same vlans.

Create new netlink support functions to delete interfaces from a trunk
and get a complete list of bridge interfaces and their netlink vlan info.