Return success if plugin config is not found in CmdDel
If cmdDel() fails, cached netconf is cleaned up by
the followed defer call or might not exist in the first place.
However, subsequence calls of cmdDel()
from container runtime fail in a dead loop because
the cached netconf doesn't exist.
Return nil when cache.Load() fails since the rest
of cmdDel() code relies on netconf as input argument
and there is no meaning to continue.

Signed-off-by: amaslennikov <amaslennikov@nvidia.com>